% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python-env.R
\name{.ensure_croberta_env}
\alias{.ensure_croberta_env}
\title{Ensure python environment for croberta scoring}
\usage{
.ensure_croberta_env(envname = "hf-env")
}
\description{
Ensure python environment for croberta scoring
}
\keyword{internal}
